Update bookSchema.json
Added to bookSchema to permit a list of files in the json with filename and optional visible (default false). Assuming that any .py will be run as pre-code and any other file type would not. It doesn't error out in local testing but when I try to add a files element to the json it still removes it as invalid.