added simple turtle/canvas automark based on matching final canvas
For discussion/demo purposes.
See new turtle automark example in the example book.
It matches the canvas output but currently requires long base64 canvas representation within the book.json for the test output. It also requires the user to have produced the canvas output before submitting but that's prob. ok