Updating to store addFiles in Challenge
Good step in the right direction I think then - still a bit more to do
- Challenge/ChallengeEditor now keeps additionalFilesLoaded (ie the contents once fetched) in its state
- in addition to the additionalFileProperties within its bookNode which is the JSON with the file details
These are written into python on each debug which works but currently it does this by adding extra lines of code before the code that is to be executed so any python reported error numbers are out e.g. error on line 3 might actually mean error on line 1 because 2 lines of file insertion have been added.