Improve online editor (#64)
Adding the new "Solutions" tab, so you can specify a model answer for your challenges. These are not visible to students.
Adding the ability to check tests against model answers. This is helpful to verify model answers and to ensure that tests are correct
For server-enabled deployments, making it easier to open the editor and to upload the currently edited book to the server.